+++ src/mySql/selectForUpdateSql.js
@@ -1,3 +1,10 @@
-module.exports = function(_alias) {
-	return ' FOR UPDATE';
+module.exports = function(_context, lock) {
+	if (typeof lock === 'string')
+		lock = { forUpdate: true };
+	let sql = '';
+	if (lock.forUpdate)
+		sql += ' FOR UPDATE';
+	if (lock.skipLocked)
+		sql += ' SKIP LOCKED';
+	return sql;
 };