+++ src/oracle/selectForUpdateSql.js
@@ -1,5 +1,10 @@
-const quote = require('../table/quote');
-
-module.exports = function(alias) {
-	return ' FOR UPDATE OF ' + quote(alias);
+module.exports = function(_context, lock) {
+	if (typeof lock === 'string')
+		lock = { forUpdate: true };
+	let sql = '';
+	if (lock.forUpdate)
+		sql = ' FOR UPDATE';
+	if (lock.skipLocked)
+		sql += ' SKIP LOCKED';
+	return sql;
 };