+++ docs/changelog.md
@@ -1,4 +1,10 @@
 ## Changelog
+__5.4.2__
+Support for limit and orderBy in `aggregate` and `distinct`
+__5.4.1__
+Fix dateWithTimeZone() column (PostgreSQL timestamptz) decodes to a non-standard string. [#195](https://github.com/alfateam/orange-orm/issues/195)
+__5.4.0__
+Support for `Select for update` and `skip locked`. [#193](https://github.com/alfateam/orange-orm/issues/193)
 __5.3.6__
 Fix npm package contents by excluding the demo project and local install artifacts.
 __5.3.5__